No extra timestamps, instead a bool

+++ CommonDataFramework/Modules/PedDatabase/PedDataController.cs
@@ -56,6 +56,14 @@ public static class PedDataController
         foreach (KeyValuePair<Ped, PedData> entry in Database.ToArray())
         {
             if (entry.Key.Exists()) continue;
+
+            // Let the data live until the next prune.
+            if (!entry.Value.RemoveDuringNextPrune)
+            {
+                entry.Value.RemoveDuringNextPrune = true;
+                continue;
+            }
+            
             Database.Remove(entry.Key);
             removed++;
         }
